# src/controllers/categories.controller.ts
import { Inject, Service } from "typedi"
import { Controller, Delete, Get, Post, Put } from "../core"
import { CategoryService } from "../services/categories.service"
import { CategoryUpadteValidation, CategoryValidation } from "../validations/category.validation";
import { NextFunction, Request, Response } from 'express';
import { ObjectId } from "mongodb";
import { Category } from "../entities/categories.entity";
import ApiError from "../utils/apiError";
import { ShapeService } from "../services/shape.service";
import { FilterUtils } from "../utils/filterUtils";


@Service()
@Controller('/categories')
export default class CategoriesController {
    @Inject(() => CategoryService)
    private readonly categoryService: CategoryService

    @Inject(() => ShapeService)
    private readonly shapeService: ShapeService


    @Post('/', CategoryValidation, {
        authorizedRole: 'all',
        isAuthenticated: false
    },
        Category)
    async createCategory(req: Request, res: Response, next: NextFunction): Promise<void> {
        try {
            const { parent, name } = req.body;
            let path: string = name;
            let ancestors: ObjectId[] = []
            if (parent) {
                const catPath = await this.categoryService.getCategoryPath(parent, name)
                path = catPath.path;
                ancestors = catPath.ancestors
            }
            const newShape = await this.categoryService.create({ ...req.body, path, ancestors, ...(parent && { parent: new ObjectId(parent) }) });
            res.status(201).json(newShape);
        } catch (e) {
            next(e)
        }
    }


    @Delete('/:id', {
        isAuthenticated: true,
        authorizedRole: 'all'
    },
        {})
    async deleteCategoryById(req: Request, res: Response, next: NextFunction): Promise<void> {
        try {
            const { categoryIdForExistingShapes } = req.query

            const category = await this.categoryService.findOneById(req.params.id);
            if (!category) {
                throw new ApiError('Category not found', 404)
            }

            const shapesCountInCategory = await this.shapeService.getCount({ category: new ObjectId(req.params.id) })
            if (shapesCountInCategory > 0) {
                if (!categoryIdForExistingShapes) {
                    throw new ApiError('categoryIdForExistingShapes for moving existing shape is required!', 400)
                }
                const categoryForShapes = await this.categoryService.findOneById(categoryIdForExistingShapes as string)
                if (!categoryForShapes) {
                    throw new ApiError('Shape Category not found', 404)
                }
                await this.shapeService.updateMany({ category: new ObjectId(req.params.id) }, { category: new ObjectId(categoryIdForExistingShapes as string) })
            }
            await this.categoryService.delete(req.params.id);
            res.status(200).json({ message: 'Category deleted successfully' });
        } catch (e) {
            next(e)
        }

    }


    @Put('/:id', CategoryUpadteValidation, {
        authorizedRole: 'all',
        isAuthenticated: false
    },
        Category || null)
    async updateCategory(req: Request, res: Response, next: NextFunction): Promise<void> {
        try {
            const categoryId = req.params.id
            const { parent, name } = req.body;
            let reqBody = { ...req.body, ...(parent && { parent: new ObjectId(parent) }) }
            if (parent) {
                const catPath = await this.categoryService.getPathOnParentChange(categoryId, parent, name)
                reqBody = { ...reqBody, path: catPath.path, ancestors: catPath.ancestors }
            }
            const newShape = await this.categoryService.update(categoryId, { ...reqBody, ...(parent && { parent: new ObjectId(parent) }) })
            res.status(201).json(newShape);
        } catch (e) {
            next(e)
        }
    }

    @Get('/', {
        authorizedRole: 'all',
        isAuthenticated: false
    },
        Array<Category>)
    async getCategories(req: Request, res: Response, next: NextFunction): Promise<void> {
        try {
            const { format, page = 1, limit = 10, sortName = 'createdAt', sortOrder = 'asc', ...restQuery } = req.query;

            let shapes
            if (format === 'nested') {
                shapes = await this.categoryService.getCategoriesNested()
            } else if (format === 'formatted') {
                shapes = await this.categoryService.getCategoriesFlat()
            } else {
                const allowedFields: (keyof Category)[] = ['name', 'parent', 'path'];

                const filters = FilterUtils.buildMongoFilters<Category>(restQuery, allowedFields);

                const sort: Record<string, 1 | -1> = { [sortName as string]: sortOrder === 'asc' ? 1 : -1 };
                shapes = await this.shapeService.findWithFilters(filters, parseInt(page as string, 10), parseInt(limit as string, 10), sort);
            }
            res.status(201).json(shapes);
        } catch (e) {
            next(e)
        }
    }

    @Get('/:id', {
        isAuthenticated: true,
        authorizedRole: 'all'
    },
        Category)
    async getCategoryById(req: Request, res: Response, next: NextFunction): Promise<void> {
        try {
            const category = await this.categoryService.findOneById(req.params.id);
            if (!category) {
                throw new ApiError('Category not found', 404)
            }
            res.status(200).json(category);
        } catch (e) {
            next(e)
        }

    }
}