# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Shapes and Categories API",
    "version": "1.0.0",
    "description": "API documentation for managing categories and shapes."
  },
  "paths": {
    "/categories": {
      "get": {
        "summary": "Get all categories with optional formatting and filtering",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "description": "Optional. Specifies the response format. By default, the response is formatted. Use 'nested' for a nested category structure.filters and sorting will not work if any format is selected",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "formatted",
                "nested"
              ],
              "default": "formatted"
            },
            "example": "formatted"
          },
          {
            "name": "sortName",
            "in": "query",
            "description": "Optional. The field to sort by. Defaults to 'createdAt'.",
            "required": false,
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "example": "createdAt"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "description": "Optional. The sort order, either 'asc' for ascending or 'desc' for descending. Defaults to 'asc'.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            },
            "example": "asc"
          },
          {
            "name": "filters",
            "in": "query",
            "description": "Dynamic filters for querying categories. Use dot notation to specify filters (e.g., `filters[name][$operator]=value`). The following fields are supported for filtering: **`name`**, **`parent`**, **`path`**. Allowed operators: `$eq`, `$ne`, `$in`, `$gte`, `$lte`, `$gt`, `$lt`.",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "examples": {
              "Example 1": {
                "value": {
                  "filters[name][$eq]": "default parent",
                  "filters[parent][$ne]": "null",
                  "filters[path][$in]": "default parent"
                }
              },
              "Example 2": {
                "value": {
                  "filters[name][$gte]": "layer shape 3",
                  "filters[path][$lt]": "path2"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of formatted categories or nested categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  },
                  "example": [
                    {
                      "_id": "677280b9db9875cbb8a7d301",
                      "createdAt": "2024-12-30T11:15:05.294Z",
                      "updatedAt": "2024-12-30T11:15:05.294Z",
                      "status": "active",
                      "name": "default parent",
                      "description": "new category for accion labs child",
                      "parent": null,
                      "path": "default parent",
                      "metadata": null,
                      "ancestors": [],
                      "allDescendants": [
                        {
                          "_id": "6773a713db9875cbb8a7d302",
                          "createdAt": "2024-12-31T08:10:59.911Z",
                          "updatedAt": "2024-12-31T08:10:59.911Z",
                          "status": "active",
                          "name": "default parent",
                          "description": "new category for accion labs child",
                          "parent": "677280b9db9875cbb8a7d301",
                          "path": "default parent/default parent",
                          "metadata": null,
                          "ancestors": [
                            "677280b9db9875cbb8a7d301"
                          ],
                          "level": 0
                        }
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "note": "Filters will work when no format is selected. Filters can be applied to fields like 'name', 'parent', and 'path' using the following operators: '$eq', '$ne', '$in', '$gte', '$lte', '$gt', '$lt'. Example: categories?filters[name][$eq]=layer shape 3"
      },
      "post": {
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              },
              "example": {
                "name": "default parent",
                "description": "new category for accion labs",
                "parent": null
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "summary": "Update a category by ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              },
              "example": {
                "name": "default parent child v1",
                "description": "new category for accion labs child",
                "parent": "677280b9db9875cbb8a7d301",
                "Metadata": {
                  "displayOrder": "1",
                  "icon": "default icon",
                  "customProperties": {
                    "name": "default parent",
                    "value": "small"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get category by ID",
        "description": "Retrieve a specific category by its unique ID.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the category",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "12345abcde"
                        },
                        "name": {
                          "type": "string",
                          "example": "Category Name"
                        },
                        "description": {
                          "type": "string",
                          "example": "Detailed description of the category"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-31T12:00:00Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-31T12:00:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Category not found"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid ID format"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a category by ID",
        "description": "Deletes a category based on the provided ID. The ID must be a valid ObjectId. Additionally, you can pass `categoryIdForExistingShapes` as a query parameter to associate the deletion with shapes under that category.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the category to delete",
            "schema": {
              "type": "string",
              "example": "677cd8973214ab562b333a52"
            }
          },
          {
            "name": "categoryIdForExistingShapes",
            "in": "query",
            "required": false,
            "description": "The ID of the category for existing shapes to be affected by the deletion",
            "schema": {
              "type": "string",
              "example": "677baa5b56b727fad9a71cd8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Category deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid ID format"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Category not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Something went wrong"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "parent": {
            "type": "string",
            "description": "if parent id given then its a child, otherwise optional"
          },
          "path": {
            "type": "string",
            "description": "The computed path of the category. (Read-only)",
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "properties": {
              "displayOrder": {
                "type": "integer"
              },
              "icon": {
                "type": "string"
              },
              "customProperties": {
                "type": "object"
              }
            },
            "description": "Additional metadata for the category. (Optional)"
          },
          "ancestors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": " (Read-only)",
            "readOnly": true
          }
        }
      },
...
    }
  }
}